backend/requirements.txt
# Web & Bot
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-telegram-bot>=20.7
pydantic-settings>=2.0.0

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Task Queue
celery[redis]>=5.3.0
redis>=5.0.0

# File Processing
PyMuPDF>=1.23.0
pdfplumber>=0.10.0
python-docx>=1.1.0
openpyxl>=3.1.0
Pillow>=10.0.0
pytesseract>=0.3.10

# Security & Validation
python-magic>=0.4.27
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9

# Storage
boto3>=1.34.0

# HTTP (Gotenberg client)
httpx>=0.26.0

# Monitoring
prometheus-client>=0.19.0
